Moving public methods out of REF compute file and adding descriptions
